:sparkles: feat: 장바구니 옵션 변경 response dto, vo 생성

+++ src/main/java/colorful/starbucks/cart/dto/CartProductOptionDto.java
@@ -6,11 +6,11 @@ import lombok.Getter;
 @Getter
 public class CartProductOptionDto {
 
-    private int colorId;
-    private int sizeId;
+    private Long colorId;
+    private Long sizeId;
 
     @Builder
-    public CartProductOptionDto(int colorId, int sizeId) {
+    public CartProductOptionDto(Long colorId, Long sizeId) {
         this.colorId = colorId;
         this.sizeId = sizeId;
     }
